# Faster constructor 