On top of https://github.com/gadget-inc/mobx-quick-tree/pull/65

15% improvement by removing a megamorphic call to the fast instantiator function! Yeesh!

On `perf-improvements`:

```
❯ node dist/bench/create-many-different-roots.js
instantiating a large root x 3,258 ops/sec ±0.84% (100 runs sampled)
instantiating a small root x 10,573,575 ops/sec ±0.21% (92 runs sampled)
instantiating a diverse root x 464,002 ops/sec ±0.15% (100 runs sampled)
```

On this branch:

```
❯ node dist/bench/create-many-different-roots.js
instantiating a large root x 3,851 ops/sec ±0.29% (97 runs sampled)
instantiating a small root x 15,595,988 ops/sec ±0.46% (99 runs sampled)
instantiating a diverse root x 582,571 ops/sec ±0.65% (99 runs sampled)
```